pyproject.toml
[tool.poetry]
name = "TribeBot - Designed for Mr. Brown (Discord)"
version = "0.1.0"
description = "A custom bot that sorts users into Tribes after completing a quiz, then assigning them the correct tribe role"
authors = ["DLCHAMP <36091350+dlchamp@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
disnake = "^2.5.2"
black = "^22.6.0"
loguru = "^0.6.0"
python-dotenv = "^0.20.0"
boto3 = "^1.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
